requirements.txt
# Core
torch>=2.2
numpy>=1.26
scipy>=1.11

# Tokenizer
tokenizers>=0.15

# RAG
scikit-learn>=1.3
faiss-cpu>=1.8

# API / UI
fastapi>=0.110
uvicorn[standard]>=0.30
streamlit>=1.32

# Utils
tqdm>=4.66
pydantic
